Added note about the Debugger being developed; updated BufferManager.
I started creating the Debugger module quite some time back, as my
live coding project. That was set aside as work and 64bit things
took up more time. However, it is still there and should be quite
functional.

The BufferManager has been updated with the handling of the system
buffers and *command.