Updated many modules with 64bit versions that work.
The 64bit versions weren't working because the relocation tables were
not built properly. These have been rebuilt now, and the modules should
be working properly with the 0.73 version of RISC OS Pyromaniac.

However, the 0.73 hasn't been released yet, so they're going to fail
quite badly - there's a heap corruption problem that needs to be applied.