Update the phases to be more complete; generate gantt charts.
The phases are now a little more completely specified, and I've started
writing some notes on why they're that way. Only two sections are done
yet, but it's getting there. The Gantt chart generation makes some quite
useful charts that are a good starting point for people to discuss the
work needed and how it might be organised.