Fix failure to render footnotes in the Wiki.
Now that we've changed to using the Wiki to store our status, we hit
the problem that footnotes don't work in there. This is an issue
because we're reliant on them to report notes about the status
of each component. Fortunately, they can be fixed up with a bit of
replacements in perl.