Add the Convert* modules which have been built.
The Convert* modules have mostly built now. The GIF and XBM modules
fail beacuse we don't have a working export of the libraries they
are based on seems to have generated error definitions which don't
work with the RISC OS 64 linkage - the _rmf64_entry is included in
the error definition. Needs a new version of CMunge exporting.