Increase timeout for the CI testing of binaries.
CI binary testing should not take very long, but we have a lot of rms
and aifs, so each takes a while to start up. Fortunately, this is now
working quite well, so we can rely on it. 20 minutes should be ok -
yes I'm hitting the build service quite a bit, but that's what it's
there for. I've also filtered the CI test runs so that it only happens
when the changes are made to the rm and aif directories (or the tests)
so that our docs changes don't affect things.