Update the main status with language bits.
Rust is now able to be used (purely as a prototype - not even as useful
as Fortran).
TCC is noted as having been built on RISC OS.

Kernel subcomponents noted in the main status, and Modules updated to
say that it's got a stub implementation present. It's probably a little
exaggeration, but it's not *that* unreasonable.