Added the Hourglass module.
The Hourglass maker has been converted to be 64-bit compatible. This
is a little bit odd as it's got assembler code in it. The assembler
code is build by Python. The assembler is in ObjAsm format. For 64bit,
the ObjAsm is converted to GAS format by the Perl ObjAsm2GAS script.
The GAS code is then built and linked with the C code as usual. All of
which is a little convoluted but it does the job.